requirements.txt
# Playwright for Python
pytest==8.0.0
pytest-playwright==0.5.0
pytest-base-url==2.1.0
pytest-html==4.1.1
